Removed one TODO from netlify. Nice.
Updated the default templates

Added lots of TODOs to lib.rs for templating logic. Updated the blog card links template. Updated how we template the homepage. Added a section to take each of the .html blog post files and add their contents to the post template and then write it to the post.html file.

In CLI, cleaned up imports, formatted, added a lengthy section to make_site_dir, if the path exists then copy over the default templates to the site's dir, if they haven't been copied over yet.