Added blog-card-template Updated index and post templates to escape all { opening braces so tinytemplate can render them
LIB: Removed all file looping. We know read in the templates, build the context for them, set the default formatter for TinyTemplate (so it doesn't escape HTML), and then render it with the template engine. Makes it WAY easier to update/move around content, and will make it much easier to design docs/ways to edit sites.

CLI: Added an extra param to template_html fn.