CARGO: Updated pulldown, reqwest, tokio, serde, and serde_json packages.
CLI: Removed the ##Date header 2 from the initial post file create.

LIB: Removed done todos. Moved the Post page templating into the same loop as the Home page templating, since there's no reason to have 2 loops going through all the post files.

INDEX TEMPLATE: Updated the blog card's link text so it is readable in dark mode

POST TEMPLATE: Updated the anchor tags for the table of content and nav to use the same styles. Added a date subheading above the content of the post (because we no longe hardcode it no file creation in a header, just in the page's metadata). Also added CSS for tables, code blocks, block quotes, and images. Claude coming in clutch.

Lorum ipsum template: has all the markdown elements in it. Will use this as a pre-populated value for new blog posts. In the GUI this will have a "one-click-delete" feature so it's easy to cleanup but still a nice example/markdown reference to have.